Improve SerializerV2 API, fix bugs, add benchmarks and documentation
- Remove IDisposable and Flush() from ICodecWriter/ICodecReader (no-ops)
- Add lossless WriteDecimal/ReadDecimal using decimal.GetBits() encoding
- Fix backward compatibility: generated deserializers use fieldCount guards
- Fix SerializerRegistry: detect conflicting type registrations, null guard
- Rewrite benchmarks to compare V2 vs Newtonsoft.Json (Akka.NET default)
- Add deserialization benchmarks for flat and envelope scenarios
- Add Newtonsoft.Json-inside-V2-envelope interop tests
- Add source generator validation diagnostics (AKKA001-007)
- Add nested type handling and SerializerSetup generation to source generator
- Add ActorRef codec support: System property on SerializerV2, system-aware
  SerializerRegistry, integration tests with real actor path resolution
- Add docs/SOURCE_GENERATION.md documentation

+++ tests/Akka.Serialization.V2.Tests/Messages/OrderMessageSerializer.cs
@@ -40,8 +40,7 @@ public sealed class OrderMessageSerializer : SerializerV2
         writer.BeginObject(4);
         writer.WriteGuid(msg.OrderId);
         writer.WriteString(msg.CustomerId);
-        // Handle decimal as double for simplicity
-        writer.WriteDouble((double)msg.Amount);
+        writer.WriteDecimal(msg.Amount);
         writer.WriteDateTimeOffset(msg.PlacedAt);
     }
 
@@ -52,7 +51,7 @@ public sealed class OrderMessageSerializer : SerializerV2
         // Read known fields
         var orderId = reader.ReadGuid();
         var customerId = reader.ReadString() ?? string.Empty;
-        var amount = (decimal)reader.ReadDouble();
+        var amount = reader.ReadDecimal();
         var placedAt = reader.ReadDateTimeOffset();
 
         // Skip unknown trailing fields for forward compatibility
